requirements-dev.txt
pytest~=7.4.3
coverage~=7.6.4
pylint~=3.0.3
typing_extensions~=4.9.0
mypy~=1.12.0
types-requests~=2.32.0
pandas-stubs~=2.2.3
types-jsonschema~=4.23.0